Move message function to messages
